# api-specification.yaml
openapi: 3.0.0
info:
  title: Distributed Assessment Platform API
  version: 1.0.0
  description: |
    Scalable distributed assessment platform with automated notifications, reporting, and data export capabilities.
    
    ## Features
    - Role-based access control (Admin/Student)
    - Hierarchical content organization (Subject→Chapter→Quiz→Question)
    - Real-time quiz attempts with automatic scoring
    - Automated background jobs for daily reminders and monthly reports
    - Data export capabilities with CSV generation
    
    ## Authentication
    All endpoints require JWT Bearer token authentication except for login and registration.
    
  contact:
    name: Distributed Assessment Platform Support
    email: support@assessmentplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.assessmentplatform.com
    description: Production server
...
paths:
...
  /admin/chapter:
    get:
      tags: [Admin - Chapter]
      summary: Get all chapters
      description: Retrieve a list of all chapters in the system
      responses:
        '200':
          description: List of chapters retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterListResponse'
    post:
      tags: [Admin - Chapter]
      summary: Create a new chapter
      description: Create a new chapter under a subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterRequest'
            example:
              subject_id: 1
              name: Algebra
              description: Algebraic concepts
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterResponse'
  /admin/chapter/{chapter_id}:
    get:
      tags: [Admin - Chapter]
      summary: Get chapter by ID
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chapter details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterResponse'
    put:
      tags: [Admin - Chapter]
      summary: Update chapter
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterRequest'
      responses:
        '200':
          description: Chapter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterResponse'
    delete:
      tags: [Admin - Chapter]
      summary: Delete chapter
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chapter deleted successfully
  /admin/chapter/search:
    get:
      tags: [Admin - Chapter]
      summary: Search chapters
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        '200':
          description: Chapter search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterListResponse'

  # Admin - Quiz Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    ChapterRequest:
      type: object
      required:
        - name
        - subject_id
      properties:
        name:
          type: string
          maxLength: 512
        subject_id:
          type: integer
        description:
          type: string
          maxLength: 1024
    ChapterResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject_id:
          type: integer
        subject_name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChapterListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChapterResponse'

    # Quiz Schemas
...
security:
  - BearerAuth: []